Switch back again from Double to Float + increase record field size to 255 chars Fix unit test

+++ recommendations-common/src/main/java/eu/europeana/api/recommend/common/model/RecordVectors.java
@@ -13,13 +13,13 @@ public class RecordVectors  implements Serializable {
     private static final long serialVersionUID = 5388995526375071859L;
 
     private String id; // required for serializing responses from Embeddings API
-    private Double[] embedding;
+    private Float[] embedding; // we use float because Milvus v2.2. and v2.3. only allows saving and searching float (and not double)
 
     public RecordVectors() {
         // Default constructor, required for Jackson deserialization
     }
 
-    public RecordVectors(String id, Double[] embedding) {
+    public RecordVectors(String id, Float[] embedding) {
         this.id = id;
         this.embedding = embedding;
     }
@@ -28,7 +28,7 @@ public class RecordVectors  implements Serializable {
         return id;
     }
 
-    public Double[] getEmbedding() {
+    public Float[] getEmbedding() {
         return embedding;
     }
